+++ packages/settings/src/Domain/ListedAuthorSecretsData/ListedAuthorSecretsData.ts
@@ -1,0 +1,3 @@
+export type ListedAuthorSecretsData = {
+  secrets: Array<{ authorId: number, secret: string }>
+}