+++ packages/snjs/lib/application.ts
@@ -132,6 +132,7 @@ import {
   FeaturesEvent,
   SetOfflineFeaturesFunctionResponse,
 } from '@Services/features_service';
+import { TagsToFoldersMigrationApplicator } from './migrations/applicators/tags_to_folders';
 
 /** How often to automatically sync, in milliseconds */
 const DEFAULT_AUTO_SYNC_INTERVAL = 30_000;
@@ -928,6 +929,21 @@ export class SNApplication {
     return this.itemManager.isValidTagParent(parentTagUuid, childTagUuid);
   }
 
+  public hasTagsNeedingFoldersMigration(): boolean {
+    return TagsToFoldersMigrationApplicator.isApplicableToCurrentData(
+      this.itemManager
+    );
+  }
+
+  /**
+   * Migrates any tags containing a '.' character to sa chema-based heirarchy, removing
+   * the dot from the tag's title.
+   */
+  public async migrateTagsToFolders(): Promise<void> {
+    await TagsToFoldersMigrationApplicator.run(this.itemManager);
+    return this.sync();
+  }
+
   /**
    * Establishes a hierarchical relationship between two tags.
    */
@@ -992,6 +1008,7 @@ export class SNApplication {
     return this.itemManager.findOrCreateTagByTitle(title);
   }
 
+  /** Creates and returns the tag but does not run sync. Callers must perform sync. */
   public async createTagOrSmartTag(title: string): Promise<SNTag | SNSmartTag> {
     return this.itemManager.createTagOrSmartTag(title);
   }