+++ packages/snjs/lib/index.ts
@@ -11,6 +11,7 @@ export type {
   UuidString,
   ApplicationIdentifier,
   ApplicationEventPayload,
+  IconType,
 } from './types';
 
 export { SNApplicationGroup } from './application_group';
@@ -77,6 +78,7 @@ export {
 } from './services/component_manager/component_viewer';
 
 export { HistoryEntry } from '@Services/history/entries/history_entry';
+export { NoteHistoryEntry } from './services/history/entries/note_history_entry';
 export { PayloadManager } from './services/payload_manager';
 
 export type { TransactionalMutation } from './services/item_manager';
@@ -233,6 +235,7 @@ export {
 
 export { NoteViewController } from './ui/note_view_controller';
 export { NoteGroupController } from './ui/note_group_controller';
+export { IconsController } from './ui/icons_controller';
 
 /** Migrations */
 export { BaseMigration } from '@Lib/migrations/base';